Commit to #9 and #19: Refactored peer discovery and fixed crash bug in native app
